CLJS-2113 nth function produces different results from clojure when using a negative index on a sequence
Created a test to check that an error is thrown for the two cases that I noticed were different from CLJ.
Verified that both checks in the new test failed.
Changed the nth parameter validation to also disallow a negative index.
Verified that both checks in the new test now pass.